perf: Lazy load Chart.js and optimize chart data generation
- Use Next.js dynamic import to lazy load chart components
- Move Chart.js registration to separate component (loaded on demand)
- Replace multiple useEffect hooks with single useMemo for chart data
- Reduces initial bundle size by deferring Chart.js loading
